dev-requirements.txt
# (c) Copyright IBM Corp. 2020,2023
# Apache License, Version 2.0 (see https://opensource.org/licenses/Apache-2.0)
-r requirements.txt
# ansible-core==2.13.7; python_version >= '3'
# ansible-core==2.11.12; python_version < '3'
junit-xml==1.9  # To get JUnit xml report from ansible-test
pytest_mock==2.0.0
mock==3.0.5
requests_mock==1.8.0
pytest-xdist==1.34.0 # Need this to work around a problem running ansible-test
mock==3.0.5; python_version < '3'
pytest==4.6.11; python_version < '3'
pytest==7.2.0; python_version >= '3'
# Non-test requirements (python 3+ only)
pylint==2.5.3; python_version >= '3'
shellcheck-py==0.9.0.5; python_version >= '3'
rstcheck==6.1.2; python_version >= '3'
yamllint==1.32.0; python_version >= '3'
voluptuous==0.13.1; python_version >= '3'
ansible-lint==6.17.0; python_version >= '3.9'
pycodestyle==2.10.0; python_version >= '3'